% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_folders.R
\name{check_folders}
\alias{check_folders}
\title{3D Fluorescence spectrum processing}
\usage{
check_folders(first_spectrum_line = 55, excitation_spectrum_line = 22,
  excitation_start = 200, excitation_end = 590, excitation_step = 10,
  emission_start = 200, emission_end = 600, emission_step = 0.5,
  directory = getwd())
}
\arguments{
\item{first_spectrum_line}{Normally you shouldn't change this. It is the first line containing the spectrum in the .sp files, 55 by default}

\item{excitation_spectrum_line}{Normally you shouldn't change this. It is the line containing the selected excitation in the .sp files, 22 by default}

\item{excitation_start}{Default = 200. Only change if you are not using the default scan3D method. Otherwise, it is your first excitation wavelenght (nm).}

\item{excitation_end}{Default = 590. Only change if you are not using the default scan3D method. Otherwise, it is your last excitation wavelenght (nm).}

\item{excitation_step}{Default = 10. Only change if you are not using the default scan3D method. Change to 20 might be useful if you are running the fast method. Otherwise, it is your excitation step (nm).}

\item{emission_start}{Default = 200. See excitation_start for details. Your first emission wavelenght (nm).}

\item{emission_end}{Default = 600 See excitation_start for details. Your last emission wavelenght (nm).}

\item{emission_step}{Default = 0.5. See excitation_step for details.}

\item{directory}{Default is the working directory. Better not change it.}
}
\description{
This function simply checks the conformity of a folder containing Fluo3D .sp files.
}
\examples{
check_folders()
}
\keyword{3D}
\keyword{fluorescence}
