% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffra_detect.R
\name{diffra_detect}
\alias{diffra_detect}
\title{Function for internal use}
\usage{
diffra_detect(x, y, z, replace_by = 0)
}
\arguments{
\item{x}{Emission wavelenght}

\item{y}{Excitation wavelenght}

\item{z}{Intensity}

\item{replace_by}{Which value will replace eliminated values, default is 0, NA might be useful}
}
\description{
It simply detects and replace values in the spectrum based on x and y conditions
}
\examples{
diffra_detect(500,250,999) returns 0 because it is within the diffraction zone.
}
\keyword{3D}
\keyword{fluorescence}
