% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffu_detect.R
\name{diffu_detect}
\alias{diffu_detect}
\title{Function for internal use}
\usage{
diffu_detect(x, y, z, span = 16, replace_by = 0)
}
\arguments{
\item{x}{Emission wavelenght}

\item{y}{Excitation wavelenght}

\item{z}{Intensity}

\item{span}{Excitation "distance" from the x=y diffusion line that will also be eliminated}

\item{replace_by}{Which value will replace eliminated values, default is 0, NA might be useful}
}
\description{
It simply detects and replace values in the spectrum based on x and y conditions
}
\examples{
diffu_detect(350,350,999)returns 0 because it is within the diffusion zone.
}
\keyword{3D}
\keyword{fluorescence}
