% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_spectrum.R
\name{emission_spectrum}
\alias{emission_spectrum}
\title{Extraction of 2D emission spectrum, given any excitation wavelenght.}
\usage{
emission_spectrum(select_excitation = 254, data_list = data_interpol,
  emission_range = data_interpol$x, area_under_graph = F, start_emi = 200,
  end_emi = 400, export_pdf = T, export_png = F,
  directory = directory_processing, sample_name_ = sample_name)
}
\arguments{
\item{select_excitation}{Selected excitation wavelenght in nm. An emission spectrum will be generated for this specific excitation. Default is 350 for exampling purposes.}

\item{data_list}{The default value of "data_interpol" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise a list containing x=vector of emission wavelenghts, y=vector of excitation wavelenghts, z= a matrix containing the z values (intensity) where X must be the lines, Y must be the columns and Z must be the values of the table.}

\item{emission_range}{The default value of "data_interpol$x" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise it must be a vector containing the emission wavelenghts range.}

\item{area_under_graph}{TRUE or FALSE. If any calculation of area under curve will be performed.}

\item{start_emi}{Emission start if area_under_graph is TRUE.}

\item{end_emi}{Emission end if area_under_graph is TRUE.}

\item{export_pdf}{TRUE or FALSE.}

\item{export_png}{TRUE or FALSE.}

\item{directory}{Default is the working directory. Better not change it.}

\item{sample_name_}{Default is the output from processing function. Better not change it.}
}
\description{
Extraction of 2D emission spectrum, given any excitation wavelenght.
}
\examples{
setwd(XXXX) #XXX is your directory contaning the .sp files (copy and paste from the brownse and replace all the "\\" by "/").
LBE_fluo3D_processing() #include your options if you want.
emission_spectrum(420) #export pdf of a emission spectrum plot at excitation = 420 nm
}
\keyword{3D}
\keyword{fluorescence}
