% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitation_spectrum.R
\name{excitation_spectrum}
\alias{excitation_spectrum}
\title{Extraction of 2D excitation spectrum, given any emission wavelenght.}
\usage{
excitation_spectrum(select_emission = 254, data_list = data_interpol,
  excitation_range = data_interpol$y, area_under_graph = F,
  start_exci = 200, end_exci = 400, export_pdf = T, export_png = F,
  directory = directory_processing, sample_name_ = sample_name)
}
\arguments{
\item{select_emission}{Selected emission wavelenght in nm. An excitation spectrum will be generated for this specific emission.  Default is 350 for exampling purposes.}

\item{data_list}{The default value of "data_interpol" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise a list containing x=vector of emission wavelenghts, y=vector of excitation wavelenghts, z= a matrix containing the z values (intensity) where X must be the lines, Y must be the columns and Z must be the values of the table.}

\item{excitation_range}{The default value of "data_interpol$y" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise it must be a vector containing the excitation wavelenghts range.}

\item{area_under_graph}{In progress}

\item{start_exci}{In progress}

\item{end_exci}{In progress}

\item{export_pdf}{In progress}

\item{export_png}{In progress}

\item{directory}{Default is the working directory. Better not change it.}

\item{sample_name_}{Default is the output from processing function. Better not change it.}
}
\description{
Extraction of 2D excitation spectrum, given any emission wavelenght.
}
\examples{
setwd(XXXX) #XXX is your directory contaning the .sp files (copy and paste from the brownse and replace all the "\\" by "/").
LBE_fluo3D_processing() #include your options if you want.
excitation_spectrum(420) #export pdf of a excitation spectrum plot at emission = 420 nm
}
\keyword{3D}
\keyword{fluorescence}
