% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBE_fluo3D_check_folders.R
\name{LBE_fluo3D_check_folders}
\alias{LBE_fluo3D_check_folders}
\title{Checks if every subfolder is OK to pass the LBE_fluo3D_process_all() function.}
\usage{
LBE_fluo3D_check_folders(args_check = list(), all_directory = getwd())
}
\arguments{
\item{args_check}{List containing the conditions for the function chech_folders See individual help files for details (?check_folders)}

\item{all_directory}{Directory containing subfolders with .sp 3D fluorescence files. Default is the directory used by the user. Might be useful to manually specify another.}
}
\description{
Checks if every subfolder is OK to pass the LBE_fluo3D_process_all() function.
}
\examples{
# (!!!) First of all, verify the conditions hilighted in "IMPORTANT". (!!!)
> setwd("C:\\\\XXXX")  #where XXXX is your directory (don't forget to replace \\ by / or \\\\)
> LBE_check_folders()

}
\keyword{3D}
\keyword{fluorescence}
