% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBE_fluo3D_indicators.R
\name{LBE_fluo3D_indicators}
\alias{LBE_fluo3D_indicators}
\title{3D Fluorescence level plot by plotly, after processing}
\usage{
LBE_fluo3D_indicators(data_list = data_interpol, peak_ratio_AT = TRUE,
  peak_ratio_CA = TRUE, peak_ratio_CT = TRUE, Humif_index = TRUE,
  export_csv = TRUE, directory = directory_processing,
  sample_name_ = sample_name)
}
\arguments{
\item{data_list}{The default value of "data_interpol" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise a list containing x=vector of emission wavelenghts, y=vector of excitation wavelenghts, z= a matrix containing the z values (intensity) where X must be the lines, Y must be the columns and Z must be the values of the table.}

\item{peak_ratio_AT}{Boolean (TRUE or FALSE).}

\item{peak_ratio_CA}{Boolean (TRUE or FALSE).}

\item{peak_ratio_CT}{Boolean (TRUE or FALSE).}

\item{Humif_index}{Boolean (TRUE or FALSE).}

\item{export_csv}{Boolean (TRUE or FALSE).}

\item{directory}{Default is the working directory. Better not change it.}

\item{sample_name_}{Default is the output from processing function. Better not change it.}
}
\description{
After running LBE_fluo3D_processing() with your options, run this function to visualize and export your surface plot. Results (plots in pdf and/or png and/or interactive html) will be exported to a folder inside your directory called "Output_data".
}
\examples{
#In progress
}
\references{
In progress
}
\keyword{3D}
\keyword{fluorescence}
