% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBE_fluo3D_levelplot.R
\name{LBE_fluo3D_levelplot}
\alias{LBE_fluo3D_levelplot}
\title{3D Fluorescence level plot by plotly, after processing}
\usage{
LBE_fluo3D_levelplot(x = x_vector, y = y_vector, z = xy_vector,
  plot_zones_polygons = TRUE, export_html = FALSE, export_png = TRUE,
  export_pdf = TRUE, n_color = 10, color_palette = 1,
  directory = directory_processing, sample_name_ = sample_name,
  zones_polygons_ = "Default")
}
\arguments{
\item{x}{The default value of "x_vector" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise it must be a vector containing the emission wavelenghts range.}

\item{y}{The default value of "y_vector" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise it must be a vector containing the excitation wavelenghts range.}

\item{z}{The default value of "xy_vector" is assuming that "LBE_fluo3D_processing" have just been used. Otherwise you must provide a matrix containing the z values (intensity) where X must be the lines, Y must be the columns and Z must be the values of the table.}

\item{plot_zones_polygons}{Boolean (TRUE or FALSE). It allows you to chose wether you want to plot the polygons with the 7 zones or not.}

\item{export_html}{TRUE or FALSE. It allows you to chose wether you want to export a plotly html interactive file (might generate about 3 MB).}

\item{export_png}{TRUE or FALSE. It allows you to chose wether you want to export a png static plot.}

\item{export_pdf}{TRUE or FALSE. It allows you to chose wether you want to export a png static plot.}

\item{n_color}{Number of colors that will be picked up in the color palette to build the color scale. Normally it is relevant only for color scales containing several colors.}

\item{color_palette}{1 to 9. Default is 1 for "viridis" (optimized for all types color blindness and with no loss of information for black and white printing). Use 2 for "gray.colors" if you want directly a black and white plot. You can select other interesting color scales such magma (3), plasma (4), inferno (5), terrain.colors (6),topo.colors (7), cm.colors (8) and matlab.like2 (9). The last one is the same used in the references in the website of the International Humic Substances Society (2018).}

\item{directory}{Default is the working directory. Better not change it.}

\item{sample_name_}{Default is the output from processing function. Better not change it.}

\item{zones_polygons_}{Default corresponds to the 7 zones defined in Jimenez et. al 2015. Otherwise, you must provide a data frame object containing three columns: one "id" column where you identify your zones for each xy coordinate. one "x" column where you put all the x coordinates (emission wavelenght). one "y" column where you put evey y coordinate (excitation wavelenght). The points of a polygon must be ordered in a way that the polygon is "closed" and must include repete the first point as a final point.}
}
\description{
After running LBE_fluo3D_processing() with your options, run this function to visualize and export your surface plot. Results (plots in pdf and/or png and/or interactive html) will be exported to a folder inside your directory called "Output_data".
}
\examples{
> setwd(XXXX) #XXX is your directory contaning the .sp files (copy and paste from the brownse and replace all the "\\" by "/").
> LBE_fluo3D_processing() #include your options if you want.
> LBE_fluo3D_levelplot(color_palette="gray.colors") #export pdf and png with a black and white surface plot.
}
\keyword{3D}
\keyword{fluorescence}
