% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBE_fluo3D_process_all.R
\name{LBE_fluo3D_process_all}
\alias{LBE_fluo3D_process_all}
\title{Applies 3D Fluoresce processing and plotting to every subfolder within a given folder.}
\usage{
LBE_fluo3D_process_all(plot_all = TRUE, exci_spec = FALSE,
  emi_spec = FALSE, indic = TRUE, zones_calc = TRUE,
  zones_polygons = "Default", args_process = list(increase_resolution =
  FALSE, remove_difdif = TRUE), args_plot = list(plot_zones_polygons = TRUE,
  export_html = TRUE, export_png = TRUE, export_pdf = TRUE),
  args_exci = list(), args_emi = list(), args_indic = list(),
  all_directory = getwd())
}
\arguments{
\item{plot_all}{Default = TRUE. Use FALSE to avoid plotting.}

\item{exci_spec}{Default = FALSE. Use TRUE to include this function and specify emission as in excitation_spectrum()}

\item{emi_spec}{Default = FALSE. Use TRUE to include this function and specify excitation as in emission_spectrum()}

\item{indic}{Default = TRUE. Will include calculation of indicators to every spectrum.}

\item{zones_calc}{Default = TRUE. Will include calculation of volumes and barycenters of each zone to every spectrum.}

\item{args_process}{List containing the conditions for the function LBE_fluo3D_processing. See individual help files for details (?LBE_fluo3D_processing)}

\item{args_plot}{List containing the conditions for the function LBE_fluo3D_levelplot See individual help files for details (?LBE_fluo3D_levelplot)}

\item{args_exci}{List containing the conditions for the function excitation_spectrum See individual help files for details (?excitation_spectrum)}

\item{args_emi}{List containing the conditions for the function emission_spectrum See individual help files for details (?emission_spectrum)}

\item{args_indic}{List containing the conditions for the function LBE_fluo3D_indicators See individual help files for details (?LBE_fluo3D_indicators)}

\item{all_directory}{Directory containing subfolders with .sp 3D fluorescence files. Default is the directory used by the user. Might be useful to manually specify another.}
}
\description{
IMPORTANT:
}
\details{
1) EVERY SUBFOLDER MUST PRESENT THE SAME CONFIGURATION OF SPECTRA (ANALYSIS METHOD), FOR A SINGLE SPECTRA
Reminder: The default internal method results in 40 .sp files called XXX#01, XXX#02, ..., XXX#40 where XXX is the label defined by the user.

2) THE CODE WILL DETECT AND STOP IF THERE IS ANY SUBFOLDER THAT DOESN'T SATISTIFY THE CONDITION ABOVE
}
\examples{
# (!!!) First of all, verify the conditions hilighted in "IMPORTANT". (!!!)
> setwd("C:\\\\XXXX")  #where XXXX is your directory (don't forget to replace \\ by / or \\\\)
> LBE_fluo3D_process_all(exci_spec = T, emi_spec = T,
                         args_process = list(remove_difdif=F),
                         args_plot = list(plot_zones_polygons=F,color_palette=2),
                         args_exci = list(select_emission=327,export_png=T,export_pdf=F),
                         args_emi = list(select_excitation=328,export_png=F,export_pdf=T),
                         args_indic = list())
# The code above will plot the gray spectrum (color_pallette=2, see ?LBE_fluo3D_levelplot for details), for all subfolders of XXXX.
# It will not remove diffusion and diffraction peaks and outside areas (remove_difdif=F, see ?LBE_fluo3D_processing for details)
# It will also plot excitation and emission 2D spectras for 327 and 328 emission and excitation wavelenghts, respectively.
# Observe that emission 2D spectra will be exported as png while excitation will be exported as pdf (just as example).
# By default, all the available indicators will be exported into each subfolder and a summary in the master folder.

}
\keyword{3D}
\keyword{fluorescence}
