% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBE_fluo3D_processing.R
\name{LBE_fluo3D_processing}
\alias{LBE_fluo3D_processing}
\title{3D Fluorescence spectrum processing}
\usage{
LBE_fluo3D_processing(first_spectrum_line = 55,
  excitation_spectrum_line = 22, excitation_start = 200,
  excitation_end = 590, excitation_step = 10, export_3D_matrix = TRUE,
  increase_resolution = FALSE, new_excitation_step = 2,
  emission_start = 200, emission_end = 600, emission_step = 0.5,
  remove_max_height = FALSE, max_height = 800, replace_max = NA,
  remove_difdif = TRUE, replace_difdif_by = 0, directory = getwd())
}
\arguments{
\item{first_spectrum_line}{Normally you shouldn't change this. It is the first line containing the spectrum in the .sp files, 55 by default}

\item{excitation_spectrum_line}{Normally you shouldn't change this. It is the line containing the selected excitation in the .sp files, 22 by default}

\item{excitation_start}{Default = 200. Only change if you are not using the default scan3D method. Otherwise, it is your first excitation wavelenght (nm).}

\item{excitation_end}{Default = 590. Only change if you are not using the default scan3D method. Otherwise, it is your last excitation wavelenght (nm).}

\item{excitation_step}{Default = 10. Only change if you are not using the default scan3D method. Change to 20 might be useful if you are running the fast method. Otherwise, it is your excitation step (nm).}

\item{export_3D_matrix}{Default = TRUE. Accepts TRUE or FALSE. Use it to export or not the processed data matrix into csv files.}

\item{increase_resolution}{If you want a more beautiful plot chose TRUE to increase the excitation step from your original to new_excitation_step.}

\item{new_excitation_step}{See increase_resolution. No effect if increase_resolution is FALSE.}

\item{emission_start}{Default = 200. See excitation_start for details. Your first emission wavelenght (nm).}

\item{emission_end}{Default = 600 See excitation_start for details. Your last emission wavelenght (nm).}

\item{emission_step}{Default = 0.5. See excitation_step for details.}

\item{remove_max_height}{Default = FALSE. Pick up TRUE if you want to replace all values greater than max_height by replace_max.}

\item{max_height}{See remove_max_height. No effect if remove_max_height is FALSE.}

\item{replace_max}{See remove_max_height. No effect if remove_max_height is FALSE.}

\item{remove_difdif}{Default = TRUE. You can chose if you want to remove the diffusion and diffraction peaks and the outside areas from your plot.}

\item{replace_difdif_by}{Anological to replace_max. No effect if remove_difdif is FALSE.}

\item{directory}{Default is the working directory. Better not change it.}
}
\description{
This function treats your set of ".sp" files withou affecting them. You just have to set the R's working directory to your folder containing your .sp files of a 3D spectra. Results (tables) will be exported to a folder inside your directory called "Output_data" .
}
\examples{
> setwd(XXXX) #XXX is your directory contaning the .sp files (copy and paste from the brownse and replace all the "\\" by "/").
> LBE_fluo3D_processing() #include your options if you want.
}
\keyword{3D}
\keyword{fluorescence}
