% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_detect.R
\name{polygon_detect}
\alias{polygon_detect}
\title{Function for internal use}
\usage{
polygon_detect(coordinates, x, y, z, replace_by = 0)
}
\arguments{
\item{coordinates}{Polygon coordinate (MATRIX object)}

\item{x}{Emission wavelenght}

\item{y}{Excitation wavelenght}

\item{z}{Intensity}

\item{replace_by}{Which value will replace eliminated values, default is 0, NA might be useful}
}
\description{
It simply detects if a given coordinate is inside a polygon with a previous established function.
}
\keyword{3D}
\keyword{fluorescence}
