% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones_definition.R
\name{zones_definition}
\alias{zones_definition}
\title{Function for internal use.}
\usage{
zones_definition(user_option = "Default", polygon_dataframe = NA)
}
\arguments{
\item{user_option}{Default is "Default". It generates the 7 zones defined by Jimenez et a. (2015). Put anything different and set your own zones with the option below if you want.}

\item{polygon_dataframe}{a data frame object containing three columns: one "id" column where you identify your zones for each xy coordinate. one "x" column where you put all the x coordinates (emission wavelenght). one "y" column where you put evey y coordinate (excitation wavelenght). The points of a polygon must be ordered in a way that the polygon is "closed" and must include repete the first point as a final point.}
}
\description{
Definition of zones.
}
\examples{
zones_definition() # The output is the data frame containing the polygons for the default zones defined in Jimenez et. al (2015). Follow this example if you want to put yours as input.
}
\references{
In progress
}
\keyword{3D}
\keyword{fluorescence}
